refactor: move db migrations in cli start command
