+++ .env.test
@@ -6,6 +6,7 @@ POSTGRES_PORT=5433
 APPS_REPO_ID=repo-id
 APPS_REPO_URL=https://test.com/test
 REDIS_HOST=localhost
+REDIS_PASSWORD=redis
 INTERNAL_IP=localhost
 TIPI_VERSION=1
 JWT_SECRET=secret
