+++ packages/cli/package.json
@@ -43,6 +43,7 @@
     "vitest": "^0.32.2"
   },
   "dependencies": {
+    "@runtipi/postgres-migrations": "^5.3.0",
     "@runtipi/shared": "workspace:^",
     "axios": "^1.4.0",
     "boxen": "^7.1.1",
@@ -53,6 +54,7 @@
     "commander": "^11.0.0",
     "dotenv": "^16.3.1",
     "log-update": "^5.0.1",
+    "pg": "^8.11.1",
     "semver": "^7.5.3",
     "systeminformation": "^5.18.7",
     "web-push": "^3.6.3",
